# backend/internal/handlers/menu_photos.go
package handlers

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"os"
	"path/filepath"
	"strconv"
	"strings"
	"time"

	"github.com/google/uuid"
	"github.com/gorilla/mux"
	"github.com/nomdb/backend/internal/database"
	"github.com/nomdb/backend/internal/models"
	"github.com/nomdb/backend/internal/services"
)

const (
	maxUploadSize    = 5 << 20 // 5MB
	uploadsDir       = "./uploads/menu_photos"
	thumbnailsSubdir = "thumbnails"
)

func init() {
	// Create uploads directory if it doesn't exist (fallback for local storage)
	os.MkdirAll(uploadsDir, 0755)
}

// @Summary Get menu photos for a restaurant
// @Description Retrieve all menu photos for a specific restaurant with presigned URLs
// @Tags Photos
// @Accept json
// @Produce json
// @Param restaurantId path int true "Restaurant ID"
// @Success 200 {array} models.MenuPhoto "List of menu photos"
// @Failure 400 {object} map[string]string "Invalid restaurant ID"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /restaurants/{restaurantId}/photos [get]
func GetMenuPhotos(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	restaurantID, err := strconv.Atoi(vars["restaurantId"])
	if err != nil {
		http.Error(w, "Invalid restaurant ID", http.StatusBadRequest)
		return
	}

	ctx := context.Background()
	rows, err := database.GetPool().Query(ctx,
		`SELECT id, restaurant_id, filename, original_filename, caption, file_size, mime_type, created_at, updated_at
		FROM menu_photos
		WHERE restaurant_id = $1
		ORDER BY created_at DESC`, restaurantID)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	defer rows.Close()

	photos := []models.MenuPhoto{}
	s3Service := services.GetS3Service()

	for rows.Next() {
		var photo models.MenuPhoto
		if err := rows.Scan(
			&photo.ID, &photo.RestaurantID, &photo.Filename, &photo.OriginalFilename,
			&photo.Caption, &photo.FileSize, &photo.MimeType, &photo.CreatedAt, &photo.UpdatedAt,
		); err != nil {
			http.Error(w, err.Error(), http.StatusInternalServerError)
			return
		}

		// Generate URL based on storage type
		if s3Service != nil {
			// Generate presigned URL for S3 (valid for 1 hour)
			presignedURL, err := s3Service.GetPresignedURL(ctx, fmt.Sprintf("menu_photos/%s", photo.Filename), time.Hour)
			if err != nil {
				http.Error(w, fmt.Sprintf("Failed to generate URL: %v", err), http.StatusInternalServerError)
				return
			}
			photo.URL = presignedURL
		} else {
			// Use local file URL
			photo.URL = fmt.Sprintf("/api/uploads/menu_photos/%s", photo.Filename)
		}

		photos = append(photos, photo)
	}

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(photos)
}

// @Summary Upload a menu photo
// @Description Upload a menu photo for a restaurant (JPEG, PNG, or WebP, max 5MB)
// @Tags Photos
// @Accept multipart/form-data
// @Produce json
// @Param restaurantId path int true "Restaurant ID"
// @Param photo formData file true "Menu photo file"
// @Param caption formData string true "Photo caption"
// @Success 201 {object} models.UploadPhotoResponse "Uploaded photo details"
// @Failure 400 {object} map[string]string "Invalid request or file"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /restaurants/{restaurantId}/photos [post]
func UploadMenuPhoto(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	restaurantID, err := strconv.Atoi(vars["restaurantId"])
	if err != nil {
		http.Error(w, "Invalid restaurant ID", http.StatusBadRequest)
		return
	}

	// Parse multipart form
	r.Body = http.MaxBytesReader(w, r.Body, maxUploadSize)
	if err := r.ParseMultipartForm(maxUploadSize); err != nil {
		http.Error(w, "File too large", http.StatusBadRequest)
		return
	}

	// Get caption
	caption := r.FormValue("caption")
	if caption == "" {
		http.Error(w, "Caption is required", http.StatusBadRequest)
		return
	}

	// Get file
	file, header, err := r.FormFile("photo")
	if err != nil {
		http.Error(w, "Failed to read file", http.StatusBadRequest)
		return
	}
	defer file.Close()

	// Validate file size
	if header.Size > maxUploadSize {
		http.Error(w, fmt.Sprintf("File too large. Maximum size is %d MB", maxUploadSize/(1<<20)), http.StatusBadRequest)
		return
	}

	// Validate file type
	contentType := header.Header.Get("Content-Type")
	if !strings.HasPrefix(contentType, "image/") {
		http.Error(w, "Only image files are allowed", http.StatusBadRequest)
		return
	}

	// Validate specific image types
	validTypes := map[string]bool{
		"image/jpeg": true,
		"image/png":  true,
		"image/webp": true,
	}
	if !validTypes[contentType] {
		http.Error(w, "Only JPEG, PNG, and WebP images are allowed", http.StatusBadRequest)
		return
	}

	// Process image (resize, compress, generate thumbnail)
	imageProcessor := services.NewImageProcessor()
	fullImage, thumbnail, err := imageProcessor.ProcessUpload(file, header.Filename)
	if err != nil {
		http.Error(w, fmt.Sprintf("Failed to process image: %v", err), http.StatusBadRequest)
		return
	}

	// Generate unique filename (always use .jpg extension after processing)
	filename := uuid.New().String() + ".jpg"
	thumbnailFilename := uuid.New().String() + "_thumb.jpg"

	ctx := context.Background()
	s3Service := services.GetS3Service()
	var fileSize int64 = int64(len(fullImage))
	var photoURL string

	if s3Service != nil {
		// Upload full image to S3
		s3Key := fmt.Sprintf("menu_photos/%s", filename)
		_, err = s3Service.UploadFile(ctx, s3Key, bytes.NewReader(fullImage), "image/jpeg")
		if err != nil {
			http.Error(w, fmt.Sprintf("Failed to upload file to S3: %v", err), http.StatusInternalServerError)
			return
		}

		// Upload thumbnail to S3
		s3ThumbKey := fmt.Sprintf("menu_photos/thumbnails/%s", thumbnailFilename)
		_, err = s3Service.UploadFile(ctx, s3ThumbKey, bytes.NewReader(thumbnail), "image/jpeg")
		if err != nil {
			http.Error(w, fmt.Sprintf("Failed to upload thumbnail to S3: %v", err), http.StatusInternalServerError)
			return
		}

		// Generate presigned URL for immediate response
		photoURL, err = s3Service.GetPresignedURL(ctx, s3Key, time.Hour)
		if err != nil {
			http.Error(w, "Failed to generate URL", http.StatusInternalServerError)
			return
		}
	} else {
		// Fallback to local storage
		filePath := filepath.Join(uploadsDir, filename)
		if err := os.WriteFile(filePath, fullImage, 0644); err != nil {
			http.Error(w, "Failed to save file", http.StatusInternalServerError)
			return
		}

		// Save thumbnail
		thumbnailDir := filepath.Join(uploadsDir, thumbnailsSubdir)
		os.MkdirAll(thumbnailDir, 0755) // Ensure thumbnail directory exists
		thumbnailPath := filepath.Join(thumbnailDir, thumbnailFilename)
		if err := os.WriteFile(thumbnailPath, thumbnail, 0644); err != nil {
			http.Error(w, "Failed to save thumbnail", http.StatusInternalServerError)
			return
		}

		photoURL = fmt.Sprintf("/api/uploads/menu_photos/%s", filename)
	}

	// Save to database (always use image/jpeg as mime type after processing)
	var photo models.MenuPhoto
	err = database.GetPool().QueryRow(ctx,
		`INSERT INTO menu_photos (restaurant_id, filename, original_filename, caption, file_size, mime_type)
		VALUES ($1, $2, $3, $4, $5, $6)
		RETURNING id, restaurant_id, filename, original_filename, caption, file_size, mime_type, created_at, updated_at`,
		restaurantID, filename, header.Filename, caption, int(fileSize), "image/jpeg",
	).Scan(
		&photo.ID, &photo.RestaurantID, &photo.Filename, &photo.OriginalFilename,
		&photo.Caption, &photo.FileSize, &photo.MimeType, &photo.CreatedAt, &photo.UpdatedAt,
	)
	if err != nil {
		// Clean up uploaded file on database error
		if s3Service != nil {
			s3Service.DeleteFile(ctx, fmt.Sprintf("menu_photos/%s", filename))
		} else {
			os.Remove(filepath.Join(uploadsDir, filename))
		}
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	// Set URL
	photo.URL = photoURL

	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(models.UploadPhotoResponse{Photo: photo})
}

// @Summary Update photo caption
// @Description Update the caption of a menu photo
// @Tags Photos
// @Accept json
// @Produce json
// @Param id path int true "Photo ID"
// @Param caption body object{caption=string} true "Caption update request"
// @Success 200 {object} models.MenuPhoto "Updated photo"
// @Failure 400 {object} map[string]string "Invalid request"
// @Failure 404 {object} map[string]string "Photo not found"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /photos/{id} [put]
func UpdatePhotoCaption(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, err := strconv.Atoi(vars["id"])
	if err != nil {
		http.Error(w, "Invalid photo ID", http.StatusBadRequest)
		return
	}

	var req struct {
		Caption string `json:"caption"`
	}
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		http.Error(w, "Invalid request body", http.StatusBadRequest)
		return
	}

	if req.Caption == "" {
		http.Error(w, "Caption is required", http.StatusBadRequest)
		return
	}

	ctx := context.Background()
	var photo models.MenuPhoto
	err = database.GetPool().QueryRow(ctx,
		`UPDATE menu_photos SET caption = $1, updated_at = NOW()
		WHERE id = $2
		RETURNING id, restaurant_id, filename, original_filename, caption, file_size, mime_type, created_at, updated_at`,
		req.Caption, id,
	).Scan(
		&photo.ID, &photo.RestaurantID, &photo.Filename, &photo.OriginalFilename,
		&photo.Caption, &photo.FileSize, &photo.MimeType, &photo.CreatedAt, &photo.UpdatedAt,
	)
	if err != nil {
		http.Error(w, "Photo not found", http.StatusNotFound)
		return
	}

	photo.URL = fmt.Sprintf("/api/uploads/menu_photos/%s", photo.Filename)

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(photo)
}

// @Summary Delete a menu photo
// @Description Delete a menu photo by ID
// @Tags Photos
// @Accept json
// @Produce json
// @Param id path int true "Photo ID"
// @Success 204 "Photo deleted successfully"
// @Failure 400 {object} map[string]string "Invalid photo ID"
// @Failure 404 {object} map[string]string "Photo not found"
// @Failure 500 {object} map[string]string "Internal server error"
// @Router /photos/{id} [delete]
func DeleteMenuPhoto(w http.ResponseWriter, r *http.Request) {
	vars := mux.Vars(r)
	id, err := strconv.Atoi(vars["id"])
	if err != nil {
		http.Error(w, "Invalid photo ID", http.StatusBadRequest)
		return
	}

	ctx := context.Background()

	// Get filename before deleting from DB
	var filename string
	err = database.GetPool().QueryRow(ctx,
		"SELECT filename FROM menu_photos WHERE id = $1", id).Scan(&filename)
	if err != nil {
		http.Error(w, "Photo not found", http.StatusNotFound)
		return
	}

	// Delete from database
	result, err := database.GetPool().Exec(ctx,
		"DELETE FROM menu_photos WHERE id = $1", id)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	if result.RowsAffected() == 0 {
		http.Error(w, "Photo not found", http.StatusNotFound)
		return
	}

	// Delete file from storage (non-fatal if fails)
	s3Service := services.GetS3Service()
	if s3Service != nil {
		// Delete from S3
		s3Service.DeleteFile(ctx, fmt.Sprintf("menu_photos/%s", filename))
	} else {
		// Delete from local disk
		filePath := filepath.Join(uploadsDir, filename)
		os.Remove(filePath)
	}

	w.WriteHeader(http.StatusNoContent)
}